pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "difference generator"
authors = ["polovykh <polovykh@mail.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/AlexanderPolovykh/python-project-50"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
